requirements.txt
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
executing==2.0.1
frozendict==2.4.4
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
lxml==5.2.2
matplotlib-inline==0.1.7
MouseInfo==0.1.3
multitasking==0.0.11
numpy==2.0.0
pandas==2.2.2
parso==0.8.4
peewee==3.17.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pure-eval==0.2.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
pytweening==1.2.0
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.40
